+++ frontend/cypress/e2e/demo.cy.ts
@@ -6,6 +6,6 @@ describe("Flow principal", () => {
 
   it("devrait afficher la page d’accueil", () => {
     cy.visit("/");
-    cy.get("h1").contains("Le référentiel des applications");
+    cy.get("h1").should("contain", "Le référentiel des applications");
   });
 });